+++ include/recomputils.h
@@ -1,7 +1,6 @@
 #ifndef __RECOMPUTILS_H__
 #define __RECOMPUTILS_H__
 
-#include <PR/ultratypes.h>
 #include "modding.h"
 #include "stdarg.h"
 
@@ -25,17 +24,17 @@ typedef enum {
     RECOMP_HUD_FULL
 } RecompHUDRatio;
 
-RECOMP_IMPORT("*", void recomp_get_window_resolution(u32 *width, u32 *height))
+RECOMP_IMPORT("*", void recomp_get_window_resolution(unsigned long *width, unsigned long *height))
 RECOMP_IMPORT("*", RecompAspectRatio recomp_get_aspect_ratio_mode())
-RECOMP_IMPORT("*", f32 recomp_get_aspect_ratio())
+RECOMP_IMPORT("*", float recomp_get_aspect_ratio())
 RECOMP_IMPORT("*", RecompHUDRatio recomp_get_hud_ratio_mode())
 RECOMP_IMPORT("*", int recomp_get_refresh_rate())
 
 RECOMP_IMPORT("*", void recomp_error_message_box(const char *message))
 RECOMP_IMPORT("*", void recomp_exit_with_error(const char *message))
 RECOMP_IMPORT("*", void recomp_exit())
-RECOMP_IMPORT("*", f32 recomp_powf(f32 x, f32 y))
-RECOMP_IMPORT("*", u32 recomp_time_us())
+RECOMP_IMPORT("*", float recomp_powf(float x, float y))
+RECOMP_IMPORT("*", unsigned long recomp_time_us())
 
 RECOMP_IMPORT("*", int recomp_sprintf(char *s, const char *fmt, ...))
 RECOMP_IMPORT("*", int recomp_vsprintf(char *s, const char *fmt, va_list args))
